+++ src/docs/docs.go
@@ -3808,6 +3808,12 @@ const docTemplate = `{
                         "name": "id",
                         "in": "path",
                         "required": true
+                    },
+                    {
+                        "type": "boolean",
+                        "description": "Force stop the virtual machine",
+                        "name": "force",
+                        "in": "query"
                     }
                 ],
                 "responses": {
@@ -7026,7 +7032,7 @@ const docTemplate = `{
                         "BearerAuth": []
                     }
                 ],
-                "description": "This endpoint returns all the packer templates",
+                "description": "This endpoint returns all the packer templates. **DEPRECATED:** This endpoint will be deprecated in the future, please upgrade your calls to use the catalog service, see https://parallels.github.io/prl-devops-service/docs/devops/catalog/overview/",
                 "produces": [
                     "application/json"
                 ],
@@ -7034,6 +7040,7 @@ const docTemplate = `{
                     "Packer Templates"
                 ],
                 "summary": "Gets all the packer templates",
+                "deprecated": true,
                 "responses": {
                     "200": {
                         "description": "OK",
@@ -7067,7 +7074,7 @@ const docTemplate = `{
                         "BearerAuth": []
                     }
                 ],
-                "description": "This endpoint creates a packer template",
+                "description": "This endpoint creates a packer template. **DEPRECATED:** This endpoint will be deprecated in the future, please upgrade your calls to use the catalog service, see https://parallels.github.io/prl-devops-service/docs/devops/catalog/overview/",
                 "produces": [
                     "application/json"
                 ],
@@ -7075,6 +7082,7 @@ const docTemplate = `{
                     "Packer Templates"
                 ],
                 "summary": "Creates a packer template",
+                "deprecated": true,
                 "parameters": [
                     {
                         "description": "Create Packer Template Request",
@@ -7118,7 +7126,7 @@ const docTemplate = `{
                         "BearerAuth": []
                     }
                 ],
-                "description": "This endpoint returns a packer template",
+                "description": "This endpoint returns a packer template. **DEPRECATED:** This endpoint will be deprecated in the future, please upgrade your calls to use the catalog service, see https://parallels.github.io/prl-devops-service/docs/devops/catalog/overview/",
                 "produces": [
                     "application/json"
                 ],
@@ -7126,6 +7134,7 @@ const docTemplate = `{
                     "Packer Templates"
                 ],
                 "summary": "Gets a packer template",
+                "deprecated": true,
                 "parameters": [
                     {
                         "type": "string",
@@ -7165,7 +7174,7 @@ const docTemplate = `{
                         "BearerAuth": []
                     }
                 ],
-                "description": "This endpoint updates a packer template",
+                "description": "This endpoint updates a packer template. **DEPRECATED:** This endpoint will be deprecated in the future, please upgrade your calls to use the catalog service, see https://parallels.github.io/prl-devops-service/docs/devops/catalog/overview/",
                 "produces": [
                     "application/json"
                 ],
@@ -7173,6 +7182,7 @@ const docTemplate = `{
                     "Packer Templates"
                 ],
                 "summary": "Updates a packer template",
+                "deprecated": true,
                 "parameters": [
                     {
                         "description": "Update Packer Template Request",
@@ -7221,7 +7231,7 @@ const docTemplate = `{
                         "BearerAuth": []
                     }
                 ],
-                "description": "This endpoint deletes a packer template",
+                "description": "This endpoint deletes a packer template. **DEPRECATED:** This endpoint will be deprecated in the future, please upgrade your calls to use the catalog service, see https://parallels.github.io/prl-devops-service/docs/devops/catalog/overview/",
                 "produces": [
                     "application/json"
                 ],
@@ -7229,6 +7239,7 @@ const docTemplate = `{
                     "Packer Templates"
                 ],
                 "summary": "Deletes a packer template",
+                "deprecated": true,
                 "parameters": [
                     {
                         "type": "string",
@@ -10286,7 +10297,7 @@ const docTemplate = `{
 
 // SwaggerInfo holds exported Swagger Info so clients can modify it
 var SwaggerInfo = &swag.Spec{
-	Version:          "0.9.16",
+	Version:          "0.9.17",
 	Host:             "",
 	BasePath:         "/api",
 	Schemes:          []string{},